集計計算類をstats.pyへ移動．平均値の区間推定の準備．パーセント表示の有効桁数の調整．
