+++ client/match_manager/run_match.sh
@@ -54,7 +54,7 @@ opt="$opt server::game_log_dir = '$log_dir' server::text_log_dir = '$log_dir'"
 opt="$opt server::log_date_format = '%Y%m%d%H%M%S-'"
 #opt="$opt server::nr_normal_halfs = 2 server::nr_extra_halfs = 0 server::penalty_shoot_outs = false"
 opt="$opt server::nr_normal_halfs = 1 server::nr_extra_halfs = 0 server::penalty_shoot_outs = false"
-opt="$opt server::half_time = 30 server::extra_half_time = 100"
+opt="$opt server::half_time = 300 server::extra_half_time = 100"
 opt="$opt server::synch_mode = $synch_mode"
 opt="$opt server::auto_mode = true"
 opt="$opt server::team_l_start = '$team_l_start'"
